refactor: move measurement logic to dedicated module
