+++ src/influxdb.cpp
@@ -1,4 +1,6 @@
+#include "influxdb.h"
 #include "env.h"
+#include "measurement.h"
 #include "utils.h"
 
 #include <HTTPClient.h>
